# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TikTok Dashboard API",
    "description": "HTTP wrapper over the tiktok-dashboard business logic (BGM / prompts / jobs / pipeline stages). Local use only — no auth, paths are hard-coded to the developer's sibling repos.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "description": "List all jobs (newest first).",
        "operationId": "list_jobs_api_v1_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobSummary"
                  },
                  "type": "array",
                  "title": "Response List Jobs Api V1 Jobs Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create Job",
        "description": "Create a new job and persist state.json under tmp/jobs/{id}/.",
        "operationId": "create_job_api_v1_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs/{job_id}/generate": {
      "post": {
        "tags": [
          "pipeline"
        ],
        "summary": "Generate",
        "description": "Trigger Seedance clip generation for all pending clips.",
        "operationId": "generate_api_v1_jobs__job_id__generate_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StageTriggerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/whisper": {
      "post": {
        "tags": [
          "pipeline"
        ],
        "summary": "Whisper",
        "description": "Trigger Whisper subtitle generation for the merged video.",
        "operationId": "whisper_api_v1_jobs__job_id__whisper_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StageTriggerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/merge": {
      "post": {
        "tags": [
          "pipeline"
        ],
        "summary": "Merge",
        "description": "Trigger FFmpeg merge of confirmed clips with the job's BGM.",
        "operationId": "merge_api_v1_jobs__job_id__merge_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StageTriggerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobCreateRequest": {
        "properties": {
          "prompts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Prompts"
          },
          "tiktok_accounts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tiktok Accounts"
          },
          "bgm_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bgm Path"
          },
          "resolution": {
            "type": "string",
            "title": "Resolution",
            "default": "480p"
          },
          "ratio": {
            "type": "string",
            "title": "Ratio",
            "default": "9:16"
          },
          "duration": {
            "type": "integer",
            "title": "Duration",
            "default": 5
          },
          "beats_per_cut": {
            "type": "integer",
            "title": "Beats Per Cut",
            "default": 2
          },
          "bpm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bpm"
          },
          "song": {
            "type": "string",
            "title": "Song",
            "default": ""
          },
          "artist": {
            "type": "string",
            "title": "Artist",
            "default": ""
          },
          "subtitle_mode": {
            "type": "string",
            "title": "Subtitle Mode",
            "default": "whisper"
          },
          "subtitle_display": {
            "type": "string",
            "title": "Subtitle Display",
            "default": "word"
          },
          "whisper_model": {
            "type": "string",
            "title": "Whisper Model",
            "default": "medium"
          },
          "whisper_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Whisper Language"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "tiktok_bgm_song": {
            "type": "string",
            "title": "Tiktok Bgm Song",
            "default": ""
          },
          "tiktok_bgm_artist": {
            "type": "string",
            "title": "Tiktok Bgm Artist",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "prompts"
        ],
        "title": "JobCreateRequest"
      },
      "JobDetail": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "JobDetail",
        "description": "Full job state — raw JSON from state.json."
      },
      "JobSummary": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "song": {
            "type": "string",
            "title": "Song"
          },
          "artist": {
            "type": "string",
            "title": "Artist"
          },
          "overall_status": {
            "type": "string",
            "title": "Overall Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "clip_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Clip Counts"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "song",
          "artist",
          "overall_status",
          "created_at",
          "updated_at",
          "clip_counts"
        ],
        "title": "JobSummary"
      },
...
      "StageTriggerResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "started"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Stage scheduled in background. Poll GET /jobs/{job_id} for progress."
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "stage"
        ],
        "title": "StageTriggerResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}